Update BoardController와 BoardService 연결

+++ src/main/java/com/ssafy/bruteforce/service/BoardServiceImpl.java
@@ -1,6 +1,5 @@
 package com.ssafy.bruteforce.service;
 
-import java.util.List;
 import java.util.NoSuchElementException;
 
 import com.ssafy.bruteforce.dto.Answer;
@@ -18,18 +17,15 @@ public class BoardServiceImpl implements BoardService {
     @Autowired
     private BoardRepository boardDao;
 
-    // @Override
-    // public List<Question> findAllQuestions() {
-    // return boardDao.findAll();
-    // }
     @Override
     public ResultJson findAllQuestions() {
         ResultJson resultJson = new ResultJson();
         try {
             resultJson.setContents(boardDao.findAll());
         } catch (NoSuchElementException e) {
-            resultJson.setStaeFail();
+            resultJson.setStateFail();
             resultJson.setMessage("No Such Element");
+            resultJson.setContents(false);
         } catch (Exception e) {
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -37,10 +33,6 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public Question findQuestionById(String qid) {
-    // return boardDao.findById(qid).get();
-    // }
     @Override
     public ResultJson findQuestionById(String qid) {
         ResultJson resultJson = new ResultJson();
@@ -53,18 +45,15 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public List<Question> findQuestionByTitle(String title) {
-    // return boardDao.findByTitle(title);
-    // }
     @Override
     public ResultJson findQuestionByTitle(String title) {
         ResultJson resultJson = new ResultJson();
         try {
             resultJson.setContents(boardDao.findByTitle(title));
         } catch (NoSuchElementException e) {
-            resultJson.setStaeFail();
+            resultJson.setStateFail();
             resultJson.setMessage("No Such Element");
+            resultJson.setContents(false);
         } catch (Exception e) {
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -72,18 +61,15 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public List<Question> findQuestionByTag(String[] tag) {
-    // return boardDao.findByTag(tag);
-    // }
     @Override
     public ResultJson findQuestionByTag(String[] tag) {
         ResultJson resultJson = new ResultJson();
         try {
             resultJson.setContents(boardDao.findByTag(tag));
         } catch (NoSuchElementException e) {
-            resultJson.setStaeFail();
+            resultJson.setStateFail();
             resultJson.setMessage("No Such Element");
+            resultJson.setContents(false);
         } catch (Exception e) {
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -91,15 +77,12 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public void addQuestion(Question question) {
-    // boardDao.insert(question);
-    // }
     @Override
     public ResultJson addQuestion(Question question) {
         ResultJson resultJson = new ResultJson();
         try {
             boardDao.insert(question);
+            resultJson.setContents(true);
         } catch (Exception e) {
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -107,15 +90,12 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public void updateQuestion(Question question) {
-    // boardDao.save(question);
-    // }
     @Override
     public ResultJson updateQuestion(Question question) {
         ResultJson resultJson = new ResultJson();
         try {
             boardDao.save(question);
+            resultJson.setContents(true);
         } catch (Exception e) {
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -123,15 +103,19 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public void deleteQuestion(String qid) {
-    // boardDao.deleteById(qid);
-    // }
     @Override
     public ResultJson deleteQuestion(String qid) {
         ResultJson resultJson = new ResultJson();
         try {
-            boardDao.deleteById(qid);
+            boolean bcheckClosed = boardDao.findById(qid).get().isbClosed();
+            if(bcheckClosed){
+                resultJson.setStateFail();
+                resultJson.setMessage("채택되었기 때문에 삭제 불가");
+                resultJson.setContents(false);
+            }else{
+                boardDao.deleteById(qid);
+                resultJson.setContents(true);
+            }
         } catch (Exception e) {
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -139,19 +123,6 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public boolean closedUpdate(String qid, String aid) {
-    // Question getQ = boardDao.findById(qid).get();
-    // getQ.setbClosed(true);
-    // for(Answer answer:getQ.answers){
-    // if(answer.aid.equals(aid)){
-    // answer.setbSelection(true);
-    // break;
-    // }
-    // }
-    // boardDao.save(getQ);
-    // return false;
-    // }
     @Override
     public ResultJson closedUpdate(String qid, String aid) {
         ResultJson resultJson = new ResultJson();
@@ -165,6 +136,7 @@ public class BoardServiceImpl implements BoardService {
                 }
             }
             boardDao.save(getQ);
+            resultJson.setContents(true);
         } catch(Exception e){
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -172,18 +144,13 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public void addAnswer(String qid, Answer answer) {
-    // Question getQ = boardDao.findById(qid).get();
-    // getQ.answers.add(answer);
-    // boardDao.save(getQ);
-    // }
     public ResultJson addAnswer(String qid, Answer answer) {
         ResultJson resultJson = new ResultJson();
         try {
             Question getQ = boardDao.findById(qid).get();
             getQ.answers.add(answer);
             boardDao.save(getQ);
+            resultJson.setContents(true);
         } catch(Exception e){
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -191,17 +158,6 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public void updateAnswer(String qid, String aid, String contents) {
-    // Question getQ = boardDao.findById(qid).get();
-    // for(Answer answer:getQ.answers){
-    // if(answer.aid.equals(aid)){
-    // answer.setContents(contents);
-    // break;
-    // }
-    // }
-    // boardDao.save(getQ);
-    // }
     @Override
     public ResultJson updateAnswer(String qid, String aid, String contents) {
         ResultJson resultJson = new ResultJson();
@@ -214,6 +170,7 @@ public class BoardServiceImpl implements BoardService {
                 }
             }
             boardDao.save(getQ);
+            resultJson.setContents(true);
         } catch(Exception e){
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -221,17 +178,6 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public void deleteAnswer(String qid, String aid) {
-    // Question getQ = boardDao.findById(qid).get();
-    // for(Answer answer:getQ.answers){
-    // if(answer.aid.equals(aid)){
-    // getQ.answers.remove(answer);
-    // break;
-    // }
-    // }
-    // boardDao.save(getQ);
-    // }
     @Override
     public ResultJson deleteAnswer(String qid, String aid) {
         ResultJson resultJson = new ResultJson();
@@ -239,7 +185,14 @@ public class BoardServiceImpl implements BoardService {
             Question getQ = boardDao.findById(qid).get();
             for (Answer answer : getQ.answers) {
                 if (answer.aid.equals(aid)) {
-                    getQ.answers.remove(answer);
+                    if(answer.isbSelection()){
+                        resultJson.setStateFail();
+                        resultJson.setMessage("채택되었기 때문에 삭제 불가");
+                        resultJson.setContents(false);
+                    }else{
+                        getQ.answers.remove(answer);
+                        resultJson.setContents(true);
+                    }
                     break;
                 }
             }
@@ -251,12 +204,6 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public void addQuestionComment(String qid, Comment comment) {
-    // Question getQ = boardDao.findById(qid).get();
-    // getQ.comments.add(comment);
-    // boardDao.save(getQ);
-    // }
     @Override
     public ResultJson addQuestionComment(String qid, Comment comment) {
         ResultJson resultJson = new ResultJson();
@@ -264,6 +211,7 @@ public class BoardServiceImpl implements BoardService {
             Question getQ = boardDao.findById(qid).get();
             getQ.comments.add(comment);
             boardDao.save(getQ);
+            resultJson.setContents(true);
         } catch(Exception e){
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -271,17 +219,6 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public void addAnswerComment(String qid, String aid, Comment comment) {
-    // Question getQ = boardDao.findById(qid).get();
-    // for(Answer answer:getQ.answers){
-    // if(answer.aid.equals(aid)){
-    // answer.comments.add(comment);
-    // break;
-    // }
-    // }
-    // boardDao.save(getQ);
-    // }
     @Override
     public ResultJson addAnswerComment(String qid, String aid, Comment comment) {
         ResultJson resultJson = new ResultJson();
@@ -294,6 +231,7 @@ public class BoardServiceImpl implements BoardService {
                 }
             }
             boardDao.save(getQ);
+            resultJson.setContents(true);
         } catch(Exception e){
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -301,17 +239,6 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public void updateQuestionComment(String qid, String cid, String contents) {
-    // Question getQ = boardDao.findById(qid).get();
-    // for(Comment comment:getQ.comments){
-    // if(comment.cid.equals(cid)){
-    // comment.setContents(contents);
-    // break;
-    // }
-    // }
-    // boardDao.save(getQ);
-    // }
     @Override
     public ResultJson updateQuestionComment(String qid, String cid, String contents) {
         ResultJson resultJson = new ResultJson();
@@ -324,6 +251,7 @@ public class BoardServiceImpl implements BoardService {
                 }
             }
             boardDao.save(getQ);
+            resultJson.setContents(true);
         } catch(Exception e){
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -331,23 +259,6 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public void updateAnswerComment(String qid, String aid, String cid, String
-    // contents) {
-    // Question getQ = boardDao.findById(qid).get();
-    // for(Answer answer:getQ.answers){
-    // if(answer.aid.equals(aid)){
-    // for(Comment comment:answer.comments){
-    // if(comment.cid.equals(cid)){
-    // comment.setContents(contents);
-    // break;
-    // }
-    // }
-    // break;
-    // }
-    // }
-    // boardDao.save(getQ);
-    // }
     @Override
     public ResultJson updateAnswerComment(String qid, String aid, String cid, String contents) {
         ResultJson resultJson = new ResultJson();
@@ -365,6 +276,7 @@ public class BoardServiceImpl implements BoardService {
                 }
             }
             boardDao.save(getQ);
+            resultJson.setContents(true);
         } catch(Exception e){
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -372,17 +284,6 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public void deleteQuestionComment(String qid, String cid) {
-    // Question getQ = boardDao.findById(qid).get();
-    // for(Comment cmt:getQ.comments){
-    // if(cmt.cid.equals(cid)){
-    // getQ.comments.remove(cmt);
-    // break;
-    // }
-    // }
-    // boardDao.save(getQ);
-    // }
     @Override
     public ResultJson deleteQuestionComment(String qid, String cid) {
         ResultJson resultJson = new ResultJson();
@@ -395,6 +296,7 @@ public class BoardServiceImpl implements BoardService {
                 }
             }
             boardDao.save(getQ);
+            resultJson.setContents(true);
         } catch(Exception e){
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
@@ -402,22 +304,6 @@ public class BoardServiceImpl implements BoardService {
         return resultJson;
     }
 
-    // @Override
-    // public void deleteAnswerComment(String qid, String aid, String cid) {
-    // Question getQ = boardDao.findById(qid).get();
-    // for(Answer ans:getQ.answers){
-    // if(ans.aid.equals(aid)){
-    // for(Comment comment:ans.comments){
-    // if(comment.cid.equals(cid)){
-    // ans.comments.remove(comment);
-    // break;
-    // }
-    // }
-    // break;
-    // }
-    // }
-    // boardDao.save(getQ);
-    // }
     @Override
     public ResultJson deleteAnswerComment(String qid, String aid, String cid) {
         ResultJson resultJson = new ResultJson();
@@ -435,6 +321,7 @@ public class BoardServiceImpl implements BoardService {
                 }
             }
             boardDao.save(getQ);
+            resultJson.setContents(true);
         } catch(Exception e){
             resultJson.setStateUnconnect();
             resultJson.setMessage("Server Error");
